requirements.txt
APScheduler==3.11.1
greenlet==3.2.4
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
